refactor split with interests file
