+++ backend/matcha/interests.py
@@ -1,0 +1,42 @@
+from flask import Blueprint, request, jsonify
+
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_identity,
+)
+
+from matcha.app_utils import check_request_json_values
+
+from matcha.db import db_get_interests, db_set_interests
+
+bp = Blueprint("interests", __name__)
+
+
+def interests_put(id_user, request):
+    json = request.json
+
+    check_request = check_request_json_values(
+        request.headers.get("Content-Type"),
+        json,
+        ["interests"],
+    )
+
+    if check_request is not None:
+        return jsonify(check_request[0]), check_request[1]
+
+    db_set_interests(id_user, json["interests"])
+
+
+@bp.route("/api/interests", methods=("PUT", "GET"))
+@jwt_required()
+def interests():
+    id_user = get_jwt_identity()
+
+    if request.method == "PUT":
+        error_msg = interests_put(id_user, request)
+        if error_msg:
+            return error_msg
+
+    interests = db_get_interests(id_user)
+
+    return jsonify({"interests": interests}), 201
